chore: undo file renaming Signed-off-by: Sonia Sandler <ssandler@redhat.com>
