+++ packages/renderer/src/lib/statusbar/ProviderWidgetStatus.svelte
