+++ packages/renderer/src/lib/statusbar/ProviderWidgetStatusStyle.svelte
